# api.yaml
swagger: "2.0"
info:
  description: Swagger file description of Interview Calendar API
  version: "1.0.0"
  title: Interview Calendar API
consumes:
  - "application/json"
produces:
  - "application/json"
basePath: "/api/v1"
...
paths:
...
  /interview-slots/{obj_id}:
    get:
      operationId: "endpoints.interview_slot.retrieve"
      tags:
      - "interview-slots"
      summary: "Retrieve interview slots for candidate"
      description: "Retrieve interview slots for candidate matching obj_id"
      parameters:
        - name: "obj_id"
          in: "path"
          description: "Id of candidate record"
          type: "integer"
          required: True
      responses:
        200:
          description: "Successful retrieve candidates list operation"
          schema:
            type: "array"
            items:
              properties:
                start_time:
                  type: "string"
                end_time:
                  type: "string"
                scheduled_date:
                  type: "string"
                is_booked:
                  type: "boolean"
                interviewer_id:
                  type: "integer"
                interviewer:
                  type: "object"
                date_created:
                  type: "string"
                date_updated:
                  type: "string"
  /interviewers/{obj_id}/availabilities:
    get:
      operationId: "endpoints.availability_slot.retrieve"
      tags:
      - "availability-slot"
      summary: "Retrieve availability slots for interviewers"
      description: "Retrieve availability slots for interviewers matching obj_id"
      parameters:
        - name: "obj_id"
          in: "path"
          description: "Id of interviewer record"
          type: "integer"
          required: True
      responses:
        200:
          description: "Successful retrieve availability slots list operation"
          schema:
            type: "array"
            items:
              properties:
                scheduled_date:
                  type: "string"
                start_time:
                  type: "string"
                end_time:
                  type: "string"
                date_created:
                  type: "string"
                date_updated:
                  type: "string"
    post:
      operationId: "endpoints.availability_slot.create"
      tags:
      - "availability-slot"
      summary: "Create availability slot record"
      description: "Create an availability slot record matching interviewer with id - obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        description: "Id of interviewer record"
        type: "integer"
        required: True
      - name: "candidate"
        in: "body"
        schema:
          type: "object"
          properties:
            scheduled_date:
              type: "string"
            start_time:
              type: "string"
            end_time:
              type: "string"

      responses:
        200:
          description: "Successfully created an availability slot record"

    delete:
      operationId: "endpoints.availability_slot.delete"
      tags:
      - "availability-slot"
      summary: "Delete availability slot record"
      description: "Delete an availability slot record matching availability slot with id - obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        type: "string"
        required: True
      responses:
        204:
          description: "Successfully deleted an availability slot record"
  /interviewers/{obj_id}/range:
    post:
      operationId: "endpoints.availability_slot.create_range"
      tags:
      - "availability-slot"
      summary: "Create availability slot record"
      description: "Create an availability slot record matching candidate with id - obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        description: "Id of interviewer record"
        type: "integer"
        required: True
      - name: "candidate"
        in: "body"
        schema:
          type: "object"
          properties:
            start_date:
              type: "string"
            end_date:
              type: "string"
            start_time:
              type: "string"
            end_time:
              type: "string"

      responses:
        200:
          description: "Successfully created interview slot request"
  /requested/{obj_id}:
    post:
      operationId: "endpoints.requested_slot.create"
      tags:
      - "requested_slot"
      summary: "Create requested slot record"
      description: "Create an requested slot record matching candidate with id - obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        description: "Id of candidate record"
        type: "integer"
        required: True
      - name: "candidate"
        in: "body"
        schema:
          type: "object"
          properties:
            scheduled_date:
              type: "string"
            start_time:
              type: "string"
            end_time:
              type: "string"

      responses:
        200:
          description: "Successfully created interview slot request"

    delete:
      operationId: "endpoints.requested_slot.delete"
      tags:
      - "requested_slot"
      summary: "Delete requested slot record"
      description: "Delete a requested slot record matching with id matching obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        type: "string"
        required: True
      responses:
        204:
          description: "Successfully deleted a requested slot record"
  /requested/{obj_id}/range:
    post:
      operationId: "endpoints.requested_slot.create_range"
      tags:
      - "requested_slot"
      summary: "Create requested slot record"
      description: "Create an requested slot record matching candidate with id - obj_id"
      parameters:
      - name: "obj_id"
        in: "path"
        description: "Id of candidate record"
        type: "integer"
        required: True
      - name: "candidate"
        in: "body"
        schema:
          type: "object"
          properties:
            start_date:
              type: "string"
            end_date:
              type: "string"
            start_time:
              type: "string"
            end_time:
              type: "string"

      responses:
        200:
          description: "Successfully created interview slot request"