# endpoints/availability_slot.py
from datetime import timedelta

from flask import jsonify
import connexion

from app import forms, models
from services.common import InvalidUsage, success_response


def retrieve(obj_id):
    """
    retrieve availability slots matching obj_id
    :param obj_id
    :return
    """
    interviewer = models.Interviewer.query.get(obj_id)

    if not interviewer:
        raise InvalidUsage("Interviewer not found matching id - %s" % obj_id, status_code=404)

    url_params = connexion.request.args
    page = url_params.get('page', 0)
    limit = url_params.get('limit', 20)

    slots = models.AvailabilitySlot.query.filter_by(interviewer_id=interviewer.id,
                                                    is_booked=False).offset(page).limit(limit).all()

    return jsonify({"data": [i.serialize() for i in slots]})


def create(obj_id, **kwargs):
    """
    creates new interviewer record
    :param obj_id:
    :param kwargs:
    :return:
    """
    interviewer = models.Interviewer.query.get(obj_id)

    if not interviewer:
        raise InvalidUsage("Interviewer not found matching id - %s" % obj_id, status_code=404)

    request_data = kwargs.get('data')
    form = forms.AvailabilityForm(csrf_enabled=False, obj=request_data)

    if form.validate_on_submit():
        if models.AvailabilitySlot.query.filter_by(scheduled_date=form.scheduled_date.data,
                                                   start_time=form.start_time.data, end_time=form.end_time.data).first():
            raise InvalidUsage("Slot already exists", status_code=400)

        slot = models.AvailabilitySlot()
        slot.interviewer_id = interviewer.id
        form.populate_obj(slot)
        models.db.session.add(slot)
        models.db.session.commit()
        return jsonify({"message": "Availability slot on date - {} scheduled for {} - "
                                   "{}".format(slot.scheduled_date, slot.start_time, slot.end_time)})

    raise InvalidUsage(form.errors, status_code=422)


def create_range(obj_id, **kwargs):
    """
    creates new interviewer availability record for date range
    :param obj_id:
    :param kwargs:
    :return:
    """
    interviewer = models.Interviewer.query.get(obj_id)

    if not interviewer:
        raise InvalidUsage("interviewer not found matching id - %s" % obj_id, status_code=404)

    request_data = kwargs.get('data')
    form = forms.RangeForm(csrf_enabled=False, obj=request_data)

    if form.validate_on_submit():
        start_date = form.start_date.data
        end_date = form.end_date.data
        period = end_date - start_date

        for i in range(0, period.days + 1):
            select_date = start_date + timedelta(days=i)

            if models.AvailabilitySlot.query.filter_by(scheduled_date=select_date, start_time=form.start_time.data,
                                                    end_time=form.end_time.data).first():
                raise InvalidUsage("Slot already exists", status_code=400)

            slot = models.AvailabilitySlot()
            slot.interviewer_id = interviewer.id
            slot.scheduled_date = select_date
            form.populate_obj(slot)
            models.db.session.add(slot)
            models.db.session.commit()
        return jsonify({"message": "Availability slot dates created"})

    raise InvalidUsage(form.errors, status_code=422)


def delete(obj_id):
    """
    deletes availability slot record matching obj_id
    :param obj_id:
    :return:
    """
    slot = models.AvailabilitySlot.query.get(obj_id)

    if not slot:
        raise InvalidUsage("Availability slot not found matching id - %s" % obj_id, status_code=404)

    models.db.session.delete(slot)
    models.db.session.commit()

    return success_response("Interviewer successfully deleted", 204)